refactor: update to do folder structure & feat: update check to do
